Rejigged solution to better enable cleaner build of modifications.
At this point it builds and runs, but unsure whether circles model is executing correctly.

+++ code/common/visualisation/Shaders.cpp
@@ -760,7 +760,7 @@ Permenently binds a texture buffer to be loaded when useProgram() is called
 */
 int Shaders::addTextureUniform(GLuint texture, char *uniformName, GLenum type)
 {
-    GLint bufferId = textures.size();
+    GLint bufferId = (GLint)textures.size();
     GLint maxTex;
     glGetIntegerv(GL_MAX_COMBINED_TEXTURE_IMAGE_UNITS, &maxTex);
     if (bufferId<maxTex && addStaticUniform(uniformName, &bufferId))
